9_int_str_unsigned_chr_printing_done

+++ includes/ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: spark <spark@student.42.fr>                +#+  +:+       +#+        */
+/*   By: hyunja <hyunja@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/13 19:02:40 by hyunja            #+#    #+#             */
-/*   Updated: 2020/10/19 22:08:31 by spark            ###   ########.fr       */
+/*   Updated: 2020/10/20 10:25:57 by hyunja           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ typedef struct	s_set
 {
 	void		*val;
 	char		type;
-	int			num;
+	long long	num;
 	char		chr;
 	char		*va_str;
 	int			precision_yn;
